Address comments on PR #1634
* I don't understand if I should be using Docker or Docker Compose. Can
we remove the redundancy between the two, maybe supply just one of the
two files... or have one rely on the other?
https://github.com/cherrypy/cherrypy/pull/1634#pullrequestreview-146298115

Acuatlly docker-compose relies on Docker file to build the image if it
is not available on a remote registry. I'll add instructions on README
in other commit.

* I don't think it's wise to change the config for the tutorial not to use
best practices
https://github.com/cherrypy/cherrypy/pull/1634#discussion_r210143033

It is required to listen in `0.0.0.0` within the container, restricting
to localhost or IP can be done at runtime with docker.

Better would be python:3 or maybe python:3-alpine (if such a thing
exists), for future compatibility.
https://github.com/cherrypy/cherrypy/pull/1634#discussion_r210143624

`python:3-alpine` in the base image now.

Redundant entry
https://github.com/cherrypy/cherrypy/pull/1634#discussion_r210143725
https://github.com/cherrypy/cherrypy/pull/1634#discussion_r210143837

Overall `docker-compose.yml` was simplified to only the required to run.