Create a test to replicate falure. Ref #1345
Added a test that captures the failure when explicit session locking is
used with file sessions.  The session tool adds a before_finalize hook
to save the session, the FileSession class asserts that the session is
locked before saving the session.  The end result is if you use
explicit locking with file sessions and release the lock yourself an
AssertionError is raised when before_finalize runs.