Fix AssertionError with explicit session locking
Fix the AssertionError that is raised when using explicit locking with
the FileSession storage class.  The before_finalize handler from the
SessionTool calls save() on the session, which asserts that the session
is locked, which may not be the case when using explicit locking.  If
explicit locking is being used FileSession's _save() will now acquire
the lock if the session is not already locked.

Fixes #1345