Use context managers to close files properly and fix tests on PyPy
Use context managers (`with`) to ensure that all open files are closed
correctly.  This resolves resource leaks and test failures with PyPy3.7.

The code prior to this change used four approaches for closing files:

1. Using a context manager (`with` clause).

2. Using a try/finally clause.

3. Closing the file in the same scope (unreliable: file object can leak
   on exception).

4. Not closing open files at all.

The last point is a real problem for PyPy since it does not GC
unreachable objects as aggressively as CPython does.  While leaving
a function scope on CPython causes the file objects private to it to
be destroyed (and therefore closed), in PyPy they can stay dangling
for some time.  When combines with buffered writes, this means that
writes can still remain pending after returning from function.

Using a context manager is a simple, consistent way to ensure that
the file object is closed once it is no longer needed.  In turn, this
guarantees that all pending writes will be performed upon function
return and the code won't be hiting race conditions between writing
a file and reading it afterwards.